# Split encode functions, so hash and match length can be inlined.