3-5% on no assembly. 1-3% with assembly.
```
benchmark                              old ns/op     new ns/op     delta
BenchmarkEncodeDigitsDefault1e4-8      392822        385422        -1.88%
BenchmarkEncodeDigitsDefault1e5-8      5360306       5160295       -3.73%
BenchmarkEncodeDigitsDefault1e6-8      56553235      54469783      -3.68%
BenchmarkEncodeDigitsCompress1e4-8     387022        383421        -0.93%
BenchmarkEncodeDigitsCompress1e5-8     5310303       5143627       -3.14%
BenchmarkEncodeDigitsCompress1e6-8     55203155      55003143      -0.36%
BenchmarkEncodeTwainDefault1e4-8       431358        427024        -1.00%
BenchmarkEncodeTwainDefault1e5-8       5910338       5837000       -1.24%
BenchmarkEncodeTwainDefault1e6-8       61903540      61103495      -1.29%
BenchmarkEncodeTwainCompress1e4-8      432358        432358        +0.00%
BenchmarkEncodeTwainCompress1e5-8      6455369       6410366       -0.70%
BenchmarkEncodeTwainCompress1e6-8      69703990      68453915      -1.79%
```
No assembly:
```
benchmark                              old ns/op     new ns/op     delta
BenchmarkEncodeDigitsDefault1e4-8      497361        471693        -5.16%
BenchmarkEncodeDigitsDefault1e5-8      8640494       8330476       -3.59%
BenchmarkEncodeDigitsDefault1e6-8      94255390      91855255      -2.55%
BenchmarkEncodeDigitsCompress1e4-8     493028        470360        -4.60%
BenchmarkEncodeDigitsCompress1e5-8     8640494       8330476       -3.59%
BenchmarkEncodeDigitsCompress1e6-8     95105440      90905200      -4.42%
BenchmarkEncodeTwainDefault1e4-8       538697        513029        -4.76%
BenchmarkEncodeTwainDefault1e5-8       8535488       8270473       -3.10%
BenchmarkEncodeTwainDefault1e6-8       90505175      87855025      -2.93%
BenchmarkEncodeTwainCompress1e4-8      547031        519696        -5.00%
BenchmarkEncodeTwainCompress1e5-8      9820562       9325533       -5.04%
BenchmarkEncodeTwainCompress1e6-8      104805990     100805770     -3.82%
```
It is a lot of duplicated code, though.